Refactored tck tools
